createAudioElementWithLoopEvent moved to factories directory.
