+++ src/AudioPlayer.js
@@ -2,7 +2,7 @@ import React, { Component, PropTypes } from 'react';
 import arrayFindIndex from 'array-find-index';
 import classNames from 'classnames';
 
-import createAudioElementWithLoopEvent from './createAudioElementWithLoopEvent';
+import createAudioElementWithLoopEvent from './factories/createAudioElementWithLoopEvent';
 import isPlaylistValid from './utils/isPlaylistValid';
 import getDisplayText from './utils/getDisplayText';
 import getControlComponent from './utils/getControlComponent';
