+++ src/factories/createAudioElementWithLoopEvent.js
